refactor: use table and merge table and row comp.
