--- src/app/entry/entry.component.css
@@ -1,3 +1,0 @@
-.table-body-cell {
-  display: table-cell;
-}
