--- src/app/entry/entry.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { EntryComponent } from './entry.component';
-
-describe('EntryComponent', () => {
-  let component: EntryComponent;
-  let fixture: ComponentFixture<EntryComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [EntryComponent]
-    });
-    fixture = TestBed.createComponent(EntryComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
